creates util for visraam
