+++ src/js/util/visraam/index.ts
@@ -1,0 +1,3 @@
+export * from './add-visraam-class';
+export * from './get-visraam-class';
+export * from './clear-visraam-class';
